chore: move util modules to src/utl
