+++ src/utl/easing-patterns.js
