+++ src/utl/rebuildFunctionalSlots.ts
